requirements.txt
Flask==3.0.2
scapy==2.5.0
python-nmap==0.7.1
rich==13.7.1
python-dotenv==1.0.1
